requirements.txt
numpy
pillow
rembg[cpu]
tkinterdnd2
fastapi
uvicorn
python-multipart
